Fix possible data miss during stats export
Fix case when gems parented to an object with duplication setting set to
none (which is every object in the scene) will be skipped during
statistics export.
Another solution is to set dupli-face objects with id custom data on
creation, to explicitly check for it on stats export. Currently there is
only check for an object type, which still would miss gems if they
parented to a mesh object with duplication setting set to none.